Several fixes related to the internal SPF implementation:
* Build the SPF test suite if either the internal implementation or libspf2
  is selected.  Previously the tests only ran if libspf2 was selected.
  In fact the only way to test the internal implementation is to request
  libspf2 but give it a bogus location; the package falls back to the
  internal SPF implementation, but still schedules the tests.
* In several places, IPv4 addresses were being passed around as u_long
  instead of uint32_t.  On at least some environments, u_long is eight bytes
  long, which leads to false comparisons.
* When doing relative comparisons on IPv4 addresses, be sure everything
  is in host byte order.
* Re-fix the type check in the A/AAAA retrieval code.
* Don't throw away replies to the A query when T_AAAA is defined.
* Parse T_AAAA replies properly.
* Document what the CIDR function expects in terms of byte order.

+++ libopendmarc/opendmarc_internal.h
@@ -291,7 +291,7 @@ char *  	opendmarc_spf_dns_get_record(char *domain, int *reply, char *txt, size_
 int     	opendmarc_spf_dns_does_domain_exist(char *domain, int *reply);
 char *  	opendmarc_spf_dns_get_record(char *domain, int *reply, char *txt, size_t txtlen, char *cname, size_t cnamelen, int spfcheck);
 int 		opendmarc_spf_ipv6_cidr_check(char *ipv6_str, char *cidr_string);
-int 		opendmarc_spf_cidr_address(u_long ip, char *cidr_addr);
+int 		opendmarc_spf_cidr_address(uint32_t ip, char *cidr_addr);
 SPF_CTX_T *     opendmarc_spf_alloc_ctx();
 SPF_CTX_T *     opendmarc_spf_free_ctx(SPF_CTX_T *spfctx);
 int             opendmarc_spf_status_to_pass(int status, int none_pass);
